fix(app): split applications page into server wrapper + client UI to resolve RSC hook violation
